mamba install cmake

+++ docker/build_essential-ubuntu.sh
@@ -20,10 +20,16 @@ apt-get clean
 pushd $INSTALL_DIR
 
 # CMake
-curl -o cmake.tgz -L https://github.com/Kitware/CMake/releases/download/v3.25.1/cmake-3.25.1-linux-x86_64.tar.gz
-tar xzf cmake.tgz && rm cmake.tgz
-ln -s cmake-*x86_64 cmake || true
-export PATH="$PWD/cmake/bin:$PATH"
+if test -n "$(command -v mamba)" -a -n "$(command -v fix-permissions)"; then
+  mamba install -y cmake
+  mamba clean --all -f -y
+  fix-permissions "${CONDA_DIR}" /home/${NB_USER}
+else
+  curl -o cmake.tgz -L https://github.com/Kitware/CMake/releases/download/v3.25.1/cmake-3.25.1-linux-x86_64.tar.gz
+  tar xzf cmake.tgz && rm cmake.tgz
+  ln -s cmake-*x86_64 cmake || true
+  export PATH="$PWD/cmake/bin:$PATH"
+fi
 
 # ccache
 mkdir -p bin
